pyproject.toml
[project]
name = "financial-analyzer"
version = "0.1.0"
description = "Production-grade financial analysis pipeline"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pandas>=2.0",
  "yfinance>=0.2",
  "pydantic>=1.10",
  "typer[all]>=0.9",
  "sqlalchemy>=1.4",
  "pyyaml>=6.0",
]

[tool.uv]
dev-dependencies = ["ruff", "pytest"]
